requirements.txt
tensorflow>=2.10.0
flask
pillow
numpy